net: phy: micrel: fix too late link up
Since commit a4cee7207a1 ("net: phy: micrel: Add workaround for bad
autoneg") by Philipp Zabel <p.zabel@pengutronix.de>:

    Based on kernel commit d2fd719bcb0e ("net/phy: micrel: Add workaround
    for bad autoneg") by Nathan Sullivan <nathan.sullivan@ni.com>:

        Very rarely, the KSZ9031 will appear to complete autonegotiation, but
        will drop all traffic afterwards.  When this happens, the idle error
        count will read 0xFF after autonegotiation completes. Reset the PHY
        when in that state.

is it possible that we get the link to late and  barebox reports "Network is
down". The problem is that ksz9031_config_init is called again in an
error case. In this case autoneg will restart but waiting for finish is
missing.

Signed-off-by: Christian Hemp <c.hemp@phytec.de>
Tested-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>