scripts/kwboot: new parameter -n to skip a number of NAKs
On some machines the CPU resets twice and so kwboot must hit the second
window to enter debug or boot mode. For this scenario it helps to ignore
a number of NAKs. If you choose a number too high for booting, the process
is only slowed down because when the CPU enters UART boot mode it sends NAKs
when not getting any input.

This new option also helps when there are voltage fluctuations due to the
power up sequence which might be interpreted as valid chars.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>