video: ipu: register framebuffer, even when no modes are found
When no monitor is connected it might be normal that no modes are found.
Do not bail out with an error in this case, which gives an ugly error
message, but instead still register the framebuffer. This gives the
user a chance to examine it and see that the framebuffer has no modes.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>