state: Make pointing to the backend using a phandle the only supported method
All other methods are broken for some time already: When starting the
kernel the state code rewrites the state node in the device tree and
replaced the "backend" property with a phandle - even when the target
can't be described as a phandle. Since using phandles is the nicest way
to point to the storage device anyway remove the other methods.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>