state: Add working example dtsi file
This adds a working example dtsi file for the state framework. It can
be directly included by the Phytec phyFLEX i.MX6 board. The board has
been chosen because it has a wide range of different storage devices
suitable for state: SD, NAND, SPI NOR and i2c EEPROM.

This example can be used for testing and also as a template for others
to copy.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>