new make helper to decode binaries using base64
This is helpful to provide mvebu binary.0 images with a patch. When
added directly a binary patch is needed which isn't understood by
patch(1).

As it's (at least) unclear if these images are distributable in general
I don't provide a patch making use of this, but the pattern is as
follows: Add

	$(obj)/start_netgear_rn2120.pblx.kwbimg: $(board)/netgear-rn2120/binary.0

to images/Makefile.mvebu and then put a binary.0.base64 into the board
folder.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>