nv: Do not create globalvars from nvvars
When we create a new nvvar there's no need to create the corresponding
globalvar, because whoever wants to use the corresponding globalvar
will create it before usage anyway. Doing this means that we have to
test for existence of a corresponding nvvar when we create a globalvar,
and if it does, set the value of the globalvar from the nvvar.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>